pyproject.toml
[build-system] # Require setuptool version due to https://github.com/pypa/setuptools/issues/2938
requires = ["setuptools>=61.0.0", "wheel"]

[project]
name = "io4dolfinx"
version = "0.11.0.dev0"
description = "Checkpointing functionality for DOLFINx meshes/functions with ADIOS2"
authors = [{ name = "Jørgen S. Dokken", email = "dokken@simula.no" }]
license = { file = "LICENSE" }
readme = "README.md"
dependencies = ["fenics-dolfinx>=0.10.0", "packaging"]

[project.optional-dependencies]
test = [
    "pytest", 
    "coverage", 
    "ipyparallel", 
    "io4dolfinx[h5py]", 
    "io4dolfinx[pyvista]", 
    "io4dolfinx[xdmf]"
]
dev = ["pdbpp", "ipython", "mypy", "ruff"]
h5py = ["h5py"]
pyvista = ["pyvista"]
xdmf = ["h5py"]
docs = [
    "jupyter-book<2.0.0",
    "ipyparallel",
    "ipywidgets",
    "jupytext",
    "ipykernel<7.0.0",     # Note: Remove once https://github.com/ipython/ipykernel/issues/1450 is in a release
    "sphinx-codeautolink",
    "io4dolfinx[h5py]",
    "sphinx_external_toc<1.1.0",
]
all = ["io4dolfinx[test,dev,docs]"]

[tool.pytest.ini_options]
addopts = ["--import-mode=importlib"]
testpaths = ["tests"]

[tool.mypy]
ignore_missing_imports = true
# Folders to exclude
exclude = ["docs/", "build/"]
# Folder to check with mypy
files = ["src", "tests"]

[tool.ruff]
src = ["src", "tests", "docs"]
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    # Pyflakes
    "F",
    # Pycodestyle
    "E",
    "W",
    # isort
    "I001",
]


[tool.ruff.lint.isort]
known-first-party = ["io4dolfinx"]
known-third-party = [
    "basix",
    "dolfinx",
    "ffcx",
    "ufl",
    "gmsh",
    "numpy",
    "pytest",
]
section-order = [
    "future",
    "standard-library",
    "mpi",
    "third-party",
    "first-party",
    "local-folder",
]

[tool.ruff.lint.isort.sections]
"mpi" = ["mpi4py", "petsc4py"]
